scripts/requirements.txt
# Elasticsearch Stack Requirements

# Elasticsearch client
elasticsearch==8.11.0

# Deep Learning
tensorflow==2.15.0
# OR use PyTorch (comment out tensorflow):
# torch==2.1.0
# torchvision==0.16.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# HTTP requests
requests==2.31.0

# Progress bars
tqdm==4.66.1

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Async support
aiofiles==23.2.1
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
